SOC-6150: binding reports: Add binding reports. (#31)
* Binding_reports: Add sql table, entity and dao for the binding reports.
* Create GroupSpaceBindingReports :
** Add GroupSpaceBindingReport object
** Add the possibility to create/delete GroupSpaceBindingReportEntity
** Add the possibility to search GroupSpaceBindingReport from (space, groupspacebinding, group, action
** Start the create the function to get the CSV file
* create the export to csv file
* Make cleanup : remove duplicate code
* Add report binding for add action, remove action, and for update action
* change relation between space and groupBindingReport, to have delete on cascade
* Binding Report - add field stillInSpace
* use field stillPresent in binding reports csv
* Update test for testing reports and new field stillInSpace
* Binding_reports: Make Binding reports table fields internationalized.
* FixSocialBuild: Fix rest service test. (#34)
* Problem when deleting a binding
** When a group is bound to 2 spaces, and when you delete one of theses binding, the users are removed from both spaces
** This fix changes the way we get userBinding, to get only userBinding linked to this group binding, and not link to the group name
* fix comment
* Fix binding process of remove user membership synchronization.
* Add GroupSpaceBindingOperationReport model and rest entity.
* Add method getGroupSpaceBindingReportOperations.
* Add GroupSpaceBindingOperationReport model and rest entity.
* Implement rest service for retrieving binding reports.
* Fix method getGroupSpaceBindingReportOperations.
* Fix rest service.
* Add sql table for GroupSpaceBindingReportUser.
* Add GroupSpaceBindingReportUserDAO.
* Add GroupSpaceBindingReportUser model.
* Fix the case when a userBinding is removed
** If the userBinding is removed due to groupBinding remove,
** the user should stay in space if he has other binding or if he was member before
** If the userBinding is remove because user leaves the group,
** then the user should stay in space only if he has other binding
* Rename GroupSpaceBindingReportDAO.
* Add the process of saving binding reports.
* Add binding Process of retrieving binding reports operations.
* Fix GroupSpaceBindingServiceTest.
* change the function getReportsForCSV to use only one action
* comment code to make it build. I will rework on it after.
* Change object for GroupBindingReport
* fix problem in changelog, and update ReportAction and ReportUsers to use object instead of id
* update TU
* Fix method buildEntityGroupSpaceBindingReportUserFrom.
* add unit tests
* Fix retrieving process of operationsReprot.
* Fix methods updateGroupSpaceBindingReportAction and buildEntityFromGroupSpaceBindingOperationReport.
* Fix GroupSpaceBindingOperationReportEntity.
* Add space Etnity to the report entity.
* Add methods getAllFromBindingQueue.
* Complete test set
* Make sure to add binding reports from the binding queue and fix ExoSpacesAdministrationBindingReports.
* update the findReportForCSV function
* Fix method getBindingReportOperations in rest service.
* create report when the binding is created
* save the synchronize report correctly
* Fix display of binding reports table.
* make savBinding and saveReport correct
* Fix display of end date.
* Fix Display of table headers.
* Fix display of reports table columns.
* format date
* display group label & fix date display
* Add button for upload file.
* Add function uploadCSVFile.
* update report link
* Fix operation type names.
* simplify the getBindingReports
* Fix display of reports table and make sure the filter is working and code cleanup.
* Make sure the search works on space name.
* make SpaceBindingReport display sync when create or remove a binding
* Take care of delete groups
* use correct action name for export file
* update startDate and  endDate column name
* fix remove binding when user is deleted
* make filter working
* Make sure when applying binding changes to redirect to the binding reports after the erquest is fulfilled. (#40)
* when deleting a groupBinding, we want the ReportAction create directly, not when the job starts
* Fix-BindingReport-US11/13 : Fix search and filter in table binding reports. (#41)
** Fix binding reports table filter and search.
** Fix method buildEntityFromGroup.
* when 'New Binding' label to 'Add Binding'
* fix problem when a userbinding already exists,  we should not duplicate it.
* Fix-BinindgReport-US04: Make sure to display directly the report of synchronize when a boun group is removed. (#42)
* fix reports when user leave the group and action is syncrhonize
* Fix-BindingReports-UI: Update label Synchronization to Synchronize. (#43)
* Fix-BindingReports-UI-01: Fix display of binding reports table. (#47)
** Fix display of reports table filter and search field.
** Fix reports table filter.
** Fix display of reports filter.
* BindingReports_US04 : in report tab, do not count user still in space for remove lines
* Fix-BindingReports-UI-02: Fix reports table filter. (#48)
* Fix-BindingReports-UI-03: Fix alignment of space name and avatar in reports table. (#49)

+++ component/service/src/test/java/org/exoplatform/social/service/rest/GroupSpaceBindingRestServiceTest.java
@@ -16,6 +16,12 @@
  */
 package org.exoplatform.social.service.rest;
 
+import java.util.ArrayList;
+import java.util.LinkedList;
+import java.util.List;
+
+import org.junit.Test;
+
 import org.exoplatform.portal.config.UserACL;
 import org.exoplatform.services.rest.impl.ContainerResponse;
 import org.exoplatform.social.core.binding.model.GroupSpaceBinding;
@@ -26,11 +32,6 @@ import org.exoplatform.social.core.space.model.Space;
 import org.exoplatform.social.core.space.spi.SpaceService;
 import org.exoplatform.social.rest.impl.binding.GroupSpaceBindingRestResourcesV1;
 import org.exoplatform.social.service.test.AbstractResourceTest;
-import org.junit.Test;
-
-import java.util.ArrayList;
-import java.util.LinkedList;
-import java.util.List;
 
 public class GroupSpaceBindingRestServiceTest extends AbstractResourceTest {
 
@@ -62,7 +63,7 @@ public class GroupSpaceBindingRestServiceTest extends AbstractResourceTest {
     spaceService = getContainer().getComponentInstanceOfType(SpaceService.class);
     userACL = getContainer().getComponentInstanceOfType(UserACL.class);
     groupSpaceBindingService = getContainer().getComponentInstanceOfType(GroupSpaceBindingService.class);
-    groupSpaceBindingRestResourcesV1 = new GroupSpaceBindingRestResourcesV1(groupSpaceBindingService, userACL);
+    groupSpaceBindingRestResourcesV1 = new GroupSpaceBindingRestResourcesV1(spaceService, groupSpaceBindingService, userACL);
     registry(groupSpaceBindingRestResourcesV1);
   }
 
@@ -73,16 +74,15 @@ public class GroupSpaceBindingRestServiceTest extends AbstractResourceTest {
   }
 
   protected void deleteAllBindings() {
-    for (GroupSpaceBinding binding : groupSpaceBindingService.findGroupSpaceBindingsBySpace(spaceId1)) {
-      groupSpaceBindingService.deleteGroupSpaceBinding(binding);
-    }
-    for (GroupSpaceBinding binding : groupSpaceBindingService.findGroupSpaceBindingsBySpace(spaceId1)) {
-      groupSpaceBindingService.deleteGroupSpaceBinding(binding);
+    if (spaceId1 != null) {
+      for (GroupSpaceBinding binding : groupSpaceBindingService.findGroupSpaceBindingsBySpace(spaceId1)) {
+        groupSpaceBindingService.deleteGroupSpaceBinding(binding);
+      }
     }
   }
 
   @Test
-  public void testGroupspacebindings() throws Exception {
+  public void testGroupSpaceBindings() throws Exception {
 
     // Given
     startSessionAs("root");
@@ -90,61 +90,90 @@ public class GroupSpaceBindingRestServiceTest extends AbstractResourceTest {
     // Given
     List<GroupSpaceBinding> groupSpaceBindings = new LinkedList<>();
     GroupSpaceBinding binding1 = new GroupSpaceBinding();
-    binding1.setId(-1);
     binding1.setGroup("/platform/administrators");
     binding1.setSpaceId(spaceId1);
     groupSpaceBindings.add(binding1);
     tearDownbindingList.add(binding1);
 
     GroupSpaceBinding binding2 = new GroupSpaceBinding();
-    binding2.setId(-1);
     binding2.setGroup("/platform/web-contributors");
     binding2.setSpaceId(spaceId1);
     groupSpaceBindings.add(binding2);
     tearDownbindingList.add(binding2);
 
     GroupSpaceBinding binding3 = new GroupSpaceBinding();
-    binding3.setId(-1);
     binding3.setGroup("/platform/web-contributors");
     binding3.setSpaceId(spaceId1);
     groupSpaceBindings.add(binding3);
     tearDownbindingList.add(binding3);
-//    groupSpaceBindings.stream().forEach(groupSpaceBinding -> groupSpaceBindingService.saveGroupSpaceBinding(groupSpaceBinding));
+    groupSpaceBindings.stream().forEach(groupSpaceBinding -> groupSpaceBindingService.saveGroupSpaceBinding(groupSpaceBinding));
     // when
-    ContainerResponse response = service("GET",
-                                         getURLResource("groupspacebindings?spaceId=1&spaceRole=manager&limit=5&offset=0"),
-                                         "",
-                                         null,
-                                         null);
+    ContainerResponse response = service("GET", getURLResource("spaceGroupBindings/" + spaceId1), "", null, null);
     // then
-    //TODO
-//    assertEquals(200, response.getStatus());
+    assertEquals(200, response.getStatus());
 
     endSession();
   }
-
   @Test
-  public void testDeleteSpacebinding() throws Exception {
-
+  public void testGetBindingReportOperations() throws Exception {
+    
     // Given
     startSessionAs("root");
     spaceId1 = createSpace("space1", "").getId();
-
     // Given
     List<GroupSpaceBinding> groupSpaceBindings = new LinkedList<>();
     GroupSpaceBinding binding1 = new GroupSpaceBinding();
-    binding1.setId(-1);
     binding1.setGroup("/platform/administrators");
     binding1.setSpaceId(spaceId1);
     groupSpaceBindings.add(binding1);
     tearDownbindingList.add(binding1);
-//    groupSpaceBindings.stream().forEach(groupSpaceBinding -> groupSpaceBindingService.saveGroupSpaceBinding(groupSpaceBinding));
+    
+    GroupSpaceBinding binding2 = new GroupSpaceBinding();
+    binding2.setGroup("/platform/web-contributors");
+    binding2.setSpaceId(spaceId1);
+    groupSpaceBindings.add(binding2);
+    tearDownbindingList.add(binding2);
+    
+    GroupSpaceBinding binding3 = new GroupSpaceBinding();
+    binding3.setGroup("/platform/web-contributors");
+    binding3.setSpaceId(spaceId1);
+    groupSpaceBindings.add(binding3);
+    tearDownbindingList.add(binding3);
+    groupSpaceBindings.stream().forEach(groupSpaceBinding -> groupSpaceBindingService.saveGroupSpaceBinding(groupSpaceBinding));
+    // when
+    ContainerResponse response = service("GET", getURLResource("spaceGroupBindings/getBindingReportOperations"), "", null, null);
+    // then
+    assertEquals(200, response.getStatus());
+    
+    endSession();
+  }
+  
+
+  @Test
+  public void testDeleteSpaceBinding() throws Exception {
+
+    // Given
+    startSessionAs("root");
+    spaceId2 = createSpace("space2", "").getId();
+
+    // Given
+    List<GroupSpaceBinding> groupSpaceBindings = new LinkedList<>();
+    GroupSpaceBinding binding1 = new GroupSpaceBinding();
+    binding1.setGroup("/platform/administrators");
+    binding1.setSpaceId(spaceId2);
+    groupSpaceBindings.add(binding1);
+    tearDownbindingList.add(binding1);
+    GroupSpaceBinding binding = groupSpaceBindingService.saveGroupSpaceBinding(binding1);
 
     // when
-    ContainerResponse response = service("DELETE", getURLResource("groupspacebindings/" + spaceId1 + "/manager"), "", null, null);
+    ContainerResponse response = service("DELETE",
+                                         getURLResource("spaceGroupBindings/removeGroupSpaceBinding/"
+                                             + String.valueOf(binding.getId())),
+                                         "",
+                                         null,
+                                         null);
     // then
-    //TODO
-//    assertEquals(200, response.getStatus());
+    assertEquals(200, response.getStatus());
 
     endSession();
   }
